# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Interview Appointment API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/positions": {
            "get": {
                "tags": [
                    "Position"
                ],
                "summary": "Get a list of job positions",
                "operationId": "979e46c0161e2570ef5478434095cb31",
                "responses": {
                    "200": {
                        "description": "List of job positions retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Position"
                ],
                "summary": "Store a new job position",
                "operationId": "8b47544f5112eedfa8e15bfa7d23ba46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "job_type",
                                    "requirement",
                                    "description",
                                    "post_date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "job_type": {
                                        "type": "string",
                                        "example": "Full-time"
                                    },
                                    "requirement": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Bachelor's degree in Computer Science",
                                            "3+ years of experience"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Responsible for developing applications."
                                    },
                                    "post_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job position stored successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/positions/{id}": {
            "get": {
                "tags": [
                    "Position"
                ],
                "summary": "Get job position details",
                "operationId": "02987e1d2529f90862c1e6ebeed617a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job position retrieved successfully"
                    },
                    "404": {
                        "description": "Job position not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Position"
                ],
                "summary": "Update a job position",
                "operationId": "e01f80e1ad1b8b3f65a959aaf8b1f4d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "job_type",
                                    "requirement",
                                    "description",
                                    "post_date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Software Engineer"
                                    },
                                    "job_type": {
                                        "type": "string",
                                        "example": "Full-time"
                                    },
                                    "requirement": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Master's degree in Computer Science",
                                            "5+ years of experience"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Responsible for leading development teams."
                                    },
                                    "post_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job position updated successfully"
                    },
                    "404": {
                        "description": "Job position not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Position"
                ],
                "summary": "Delete a job position",
                "operationId": "be998660e7df336ea7e1b510aefe2694",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job position deleted successfully"
                    },
                    "404": {
                        "description": "Job position not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/positions/change-status/{id}": {
            "post": {
                "tags": [
                    "Position"
                ],
                "summary": "Change the status of a job position",
                "operationId": "1e0d32aa53e38c12b17c8afe144cfd33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job position status changed successfully"
                    },
                    "404": {
                        "description": "Job position not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token to access protected routes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}